% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{extract_questions}
\alias{extract_questions}
\title{Extract questions from phenotype data.}
\usage{
extract_questions(pheno_data)
}
\arguments{
\item{pheno_data}{Phenotype data.}
}
\value{
Questions.
}
\description{
Extract questions from phenotype data.
}
