% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_data}
\alias{load_data}
\title{Load survey data, phenotype data, ethnicity data.}
\usage{
load_data(survey_path, ethnicity_path, pheno_path, trait_name, default_cohorts)
}
\arguments{
\item{survey_path}{Path to survey file.}

\item{ethnicity_path}{Path to ethnicity data directory.}

\item{pheno_path}{Path to phenotypes data directory.}

\item{trait_name}{Name of the trait.}

\item{default_cohorts}{Default ethnicity cohorts.}
}
\value{
Prepared data.
}
\description{
Load survey data, phenotype data, ethnicity data.
}
