% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_ethnicity}
\alias{load_ethnicity}
\title{Load ethnicity data.}
\usage{
load_ethnicity(ethnicity_path, default_cohorts)
}
\arguments{
\item{ethnicity_path}{Path to ethnicity data directory.}

\item{default_cohorts}{Default ethnicity cohorts.}
}
\value{
Ethnicity data.
}
\description{
Load ethnicity data.
}
