% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_pheno}
\alias{load_pheno}
\title{Load phenotype data.}
\usage{
load_pheno(pheno_path, trait_name, ext = ".tsv")
}
\arguments{
\item{pheno_path}{Path to phenotypes data directory.}

\item{trait_name}{Name of the trait.}

\item{ext}{File extension (default .tsv).}
}
\value{
Phenotype data.
}
\description{
Load phenotype data.
}
