% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_survey}
\alias{load_survey}
\title{Load survey data and subset to questions of interest.}
\usage{
load_survey(survey_path, questions)
}
\arguments{
\item{survey_path}{Path to survey file.}

\item{questions}{Questions of interest.}
}
\value{
Survey data.
}
\description{
Load survey data and subset to questions of interest.
}
