% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenoreportr.R
\docType{package}
\name{phenoreportr}
\alias{phenoreportr}
\title{\code{phenoreportr}: package provides a simple R interface for cleaning and generating reports for phenotypes of interest.}
\description{
For a complete list of available functions, use \code{library(help = "phenoreportr")}.
}
\references{
Browse documentation at: \url{https://github.com/Ancestry/NatGenCOVIDpaper2022}

Browse source code at: \url{https://github.com/Ancestry/NatGenCOVIDpaper2022/tree/main/clean_pheno_package/phenoreportr}
}
