% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{phenotype_sample_size_by_response_plot}
\alias{phenotype_sample_size_by_response_plot}
\title{Generate sample size by response plot.}
\usage{
phenotype_sample_size_by_response_plot(data, questions)
}
\arguments{
\item{data}{Prepared data.}

\item{questions}{Questions.}
}
\value{
Sample size by response plot.
}
\description{
Generate sample size by response plot.
}
