% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{prevalence_plot}
\alias{prevalence_plot}
\title{Generate prevalence plot.}
\usage{
prevalence_plot(data)
}
\arguments{
\item{data}{Prepared data.}
}
\value{
Prevalence plot.
}
\description{
Generate prevalence plot.
}
