% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{prevalence_table}
\alias{prevalence_table}
\title{Generate prevalence table from prepared data.}
\usage{
prevalence_table(data)
}
\arguments{
\item{data}{Prepared data.}
}
\value{
Prevalence table.
}
\description{
Generate prevalence table from prepared data.
}
