% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{response_combined_table}
\alias{response_combined_table}
\title{Generate responses combined table from prepared data.}
\usage{
response_combined_table(data)
}
\arguments{
\item{data}{Prepared data.}
}
\value{
Responses combined table.
}
\description{
Generate responses combined table from prepared data.
}
