% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{response_sample_size_by_question_table}
\alias{response_sample_size_by_question_table}
\title{Generate response sample size by question from prepared data.}
\usage{
response_sample_size_by_question_table(data, questions)
}
\arguments{
\item{data}{Prepared data.}

\item{questions}{Questions.}
}
\value{
Response sample size by question table.
}
\description{
Generate response sample size by question from prepared data.
}
